audio-manager: Change behaviour to player-like
